Rename data classes to match google style guide
Style guide discourages the use of IMyIntergace so I changed the interface classes to **Model to show that they describe the expected data model. The data classes have the naming scheme **Data to show that they hold data. At the moment, having an interface is not necessary but I included them to to make type checking easier and so we can make test classes that inherit the interfaces