Changes to reduce repeated code made
The front and the back of the card have styling done to them.
We can see what this styling looks like in .front-content and .back-content (in our flashcard.component.scss).
Most of this styling is repeated though so it has now been placed in one section of code in that file.
The fonts are kept in seperate code sections though because the front font differs from the back font.