Imported Services into Study Flashcard Component
I imported the services into the study-flashcard component so that I can use the services such as fetching in the study-flashcard component.
I also passed the services/injected them into the constructor so that when a study-flashcard component is made, the services methods are ready to go and be used by this study-flashcard component.