Connected sign in to firebase and altered routing
I created authentication services to sign a user in. The firebase one is connected to the firebase project and allows a user to sign in through google. The routes have been changed so that homepage is the base page and that sign in links to whatever page is in the url instead of just to the homepage. The sign in still needs a way to stay active even after refreshing the site, and it may need to create a user object though this should be easy through the user service.