New additions in attempt to pass data in
I am trying to pass in the data to my enlarged flashcard based on the card that the user is on.
I am doing it the way the other methods are doing it now.
It still will not display the flashcard data but I feel that this is a step in the right direction.